refactor: introduce press section to the application standarts
