+++ apps/backoffice/src/app/public-api/knowledge-base/[slug]/route.ts
@@ -62,4 +62,3 @@ export const GET = async (_request: Request, { params }: { params: Promise<{ slu
 
 	//
 };
-
