+++ apps/backoffice/src/app/public-api/knowledge-base/route.ts
@@ -54,4 +54,3 @@ export const GET = async () => {
 
 	//
 };
-
