+++ apps/frontend/src/app/(views)/(website)/press/knowledge-base/[slug]/page.tsx
@@ -59,4 +59,3 @@ export default async function Page({ params }) {
 	const { slug } = await params;
 	return <PressKnowledgeBaseDetail slug={slug} />;
 }
-
