--- apps/frontend/src/components/home/PressHeaderGenericSection/styles.module.css
@@ -1,17 +1,0 @@
-.pressHeaderSectionWrapper {
-    height: 232px;
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-image: url('/assets/press/Header_Press.png');
-    border-radius: 18px;
-    color: var(--System-Text-100, #000);
-    text-align: center;
-    font-family: Inter;
-    font-size: 42px;
-    font-style: normal;
-    font-weight: 700;
-    line-height: normal;
-}
-
