+++ apps/frontend/src/components/press/PressGenericCard/index.tsx
