+++ apps/frontend/src/components/press/PressGenericCard/styles.module.css
