+++ apps/frontend/src/components/press/PressHeaderGenericSection/index.tsx
@@ -8,10 +8,25 @@ interface PressHeaderGenericSectionProps {
 }
 
 export function PressHeaderGenericSection({ title }: PressHeaderGenericSectionProps) {
+	// Determine title length and apply appropriate class
+	const getTitleClass = () => {
+		const titleLength = title.length;
+
+		if (titleLength <= 8) {
+			return styles.titleShort;
+		}
+		else if (titleLength <= 16) {
+			return styles.titleMedium;
+		}
+		else {
+			return styles.titleLong;
+		}
+	};
+
 	return (
 		<BreakpointDesktop>
 			<section className={styles.pressHeaderSectionWrapper}>
-				<h1>{title}</h1>
+				<h1 className={getTitleClass()}>{title}</h1>
 			</section>
 		</BreakpointDesktop>
 	);
