+++ apps/frontend/src/components/press/PressHeaderGenericSection/styles.module.css
@@ -1,0 +1,44 @@
+.pressHeaderSectionWrapper {
+    height: 232px;
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-image: url('/assets/press/Header_Press.png');
+    border-radius: 18px;
+    color: var(--System-Text-100, #000);
+    text-align: center;
+    font-family: Inter;
+    font-style: normal;
+    font-weight: 700;
+    line-height: normal;
+}
+
+/* Base title styles */
+.pressHeaderSectionWrapper h1 {
+    margin: 0;
+    padding: 0 20px;
+    max-width: 100%;
+    word-wrap: break-word;
+    hyphens: auto;
+}
+
+/* Short titles (≤15 characters) - Original size */
+.titleShort {
+    font-size: 72px;
+    line-height: 1.1;
+}
+
+/* Medium titles (16-25 characters) - Slightly smaller */
+.titleMedium {
+    font-size: 50px;
+    line-height: 1.2;
+}
+
+/* Long titles (>25 characters) - Smaller with line breaks */
+.titleLong {
+    font-size: 37px;
+    line-height: 1.3;
+    max-width: 90%;
+}
+
