+++ apps/frontend/src/components/press/PressKnowledgePage/index.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useTranslations } from 'next-intl';
 import React from 'react';
 
@@ -5,17 +7,17 @@ import styles from './styles.module.css';
 
 import { BreakpointDesktop } from '../../responsive/BreakpointSwitch';
 import { PressHeaderGenericSection } from '../PressHeaderGenericSection';
-import { PressNotesContentSection } from '../PressNotesContentSection';
+import { PressKnowledgeContentSection } from '../PressKnowledgeContentSection';
 
-export function PressNotesPage() {
-	const t = useTranslations('home.PressNotesBase');
+export function PressKnowledgePage() {
+	const t = useTranslations('press.KnowledgeBase');
 
 	return (
 		<div className={styles.container}>
 			<BreakpointDesktop>
 				<PressHeaderGenericSection title={t('section_heading')} />
 			</BreakpointDesktop>
-			<PressNotesContentSection />
+			<PressKnowledgeContentSection />
 		</div>
 	);
 }
