+++ apps/frontend/src/components/press/PressNotesPage/index.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useTranslations } from 'next-intl';
 import React from 'react';
 
@@ -5,17 +7,17 @@ import styles from './styles.module.css';
 
 import { BreakpointDesktop } from '../../responsive/BreakpointSwitch';
 import { PressHeaderGenericSection } from '../PressHeaderGenericSection';
-import { PressKnowledgeContentSection } from '../PressKnowledgeContentSection';
+import { PressNotesContentSection } from '../PressNotesContentSection';
 
-export function PressKnowledgePage() {
-	const t = useTranslations('home.PressKnowledgeBase');
+export function PressNotesPage() {
+	const t = useTranslations('press.NotesBase');
 
 	return (
 		<div className={styles.container}>
 			<BreakpointDesktop>
 				<PressHeaderGenericSection title={t('section_heading')} />
 			</BreakpointDesktop>
-			<PressKnowledgeContentSection />
+			<PressNotesContentSection />
 		</div>
 	);
 }
