+++ apps/frontend/src/components/press/PressPage/index.tsx
@@ -1,18 +1,19 @@
+'use client';
+
+import { useTranslations } from 'next-intl';
 import React from 'react';
+
+import styles from './styles.module.css';
 
 import { PressContact } from '../PressContact';
 import { PressFaqSection } from '../PressFaqSection';
 import { PressHeaderGenericSection } from '../PressHeaderGenericSection';
-
-import { PressLabelSection } from '../PressLabelSection';
-
-import styles from './styles.module.css';
-import { useTranslations } from 'next-intl';
 import { PressKnowledgeBase } from '../PressKnowledgeBase/Index';
+import { PressLabelSection } from '../PressLabelSection';
 import { PressNotesBase } from '../PressNotesBase/Index';
 
 export function PressPage() {
-	const t = useTranslations('home.PressHeaderSection');
+	const t = useTranslations('press.HeaderSection');
 	return (
 		<div className={styles.container}>
 			<PressHeaderGenericSection title={t('section_heading')} />
