+++ shared/types/src/knowledge-base.ts
@@ -80,4 +80,3 @@ export interface KnowledgeBase {
 	topic?: string
 
 }
-
